requirements.txt
# ── Web Server & API ──
fastapi>=0.104.1
python-multipart>=0.0.20
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# ── Network / Tunnel ──
pyngrok>=7.3.0

# --- Database ---
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography

# ── ML Core ──
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.26.2
transformers>=4.35.2
groq>=0.4.2

# ── WhisperX & Utilities ──
whisperx @ git+https://github.com/m-bain/whisperX.git@2d9ce44329ae73af2520196d31cd14b6192ace44
pyannote.audio>=3.1.1
imageio-ffmpeg
gTTS>=2.5.0
pydub>=0.25.1
yt-dlp
fuzzywuzzy
python-Levenshtein
jellyfish
nltk